feat: register User
