--- src/routes/auth/index.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const authController = require("../controllers/auth.Controller");
-
-router.post("/login", authController.login);
-
-module.exports = router;
