--- src/routes/controllers/auth.Controller.js
@@ -1,7 +1,0 @@
-exports.login = async (req, res, next) => {
-  try {
-    res.send({ result: "ok" });
-  } catch (err) {
-    next(err);
-  }
-};
