+++ src/routes/index.js
@@ -1,10 +1,10 @@
 const express = require("express");
 const router = express.Router();
 
-const auth = require("./auth/index");
+const user = require("./user/index");
 const assets = require("./assets/index");
 
-router.use("/auth", auth);
+router.use("/user", user);
 router.use("/assets", assets);
 
 module.exports = router;
